requirements.txt
fastapi==0.94.1
uvicorn==0.21.1
pydantic~=1.10.6
python-dotenv~=1.0.0
databases~=0.7.0
starlette~=0.26.1
SQLAlchemy~=1.4.46
asyncpg==0.26.0
passlib~=1.7.4
alembic~=1.10.2

PyJWT~=2.6.0
Authlib==1.2.0
Werkzeug==2.2.3
jose~=1.0.0